Separate stochastic base functionality from timeevolution
